blog/serializers.py
from .models import Article, category
from rest_framework import serializers


class ArticleSerializer(serializers.HyperlinkedModelSerializer):
    class Meta:
        model = Article
        fields = ('id', 'cat_id', 'title', 'content')


class CategorySerializer(serializers.HyperlinkedModelSerializer):
    class Meta:
        model = category
        fields = ('cat_id,', 'created_date', 'published_date')

