requirements.txt
awsebcli==3.14.2
botocore==1.10.50
cement==2.8.2
colorama==0.3.9
Django==2.0.6
django-ckeditor==5.5.0
django-filter==1.1.0
django-js-asset==1.1.0
djangorestframework==3.8.2
docutils==0.14
jmespath==0.9.3
Markdown==2.6.11
pathspec==0.5.5
Pillow==5.1.0
python-dateutil==2.7.3
pytz==2018.4
PyYAML==3.12
semantic-version==2.5.0
six==1.11.0
tabulate==0.7.5
termcolor==1.1.0
